refactor/atualizar-atributo-preco

+++ src/main/java/com/product/api/crud_produtos/domain/produto/dto/ProdutoAtualizarRequestDTO.java
@@ -1,10 +1,11 @@
 package com.product.api.crud_produtos.domain.produto.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProdutoAtualizarRequestDTO(
         String nome,
-        Float preco,
+        BigDecimal preco,
         Integer quantidade,
         UUID categoriaId
 ) {
